Display Invoice Rune
This commit adds an invoice rune to the ConnectWallet component.  Most values in that screen come from the backend method /shared/connectwallet and the type is found at app-config.type.ts/WalletConnect.

An invoice rune is added to WalletConnect because that is what is shown on the ConnectWallet.tsx component.  ApplicationActions.SET_WALLET_CONNECT appends an invoice rune if one is found.  If one is not found, ConnectWallet.tsx has a useEffect that checks for a missing invoice rune, and makes a createrune rpc call and then refreshes the component.